x-trainlabeling-pipeline/requirements.txt
apache-airflow==2.9.3
python-dotenv
ultralytics>=8.2.0
opencv-python
pyyaml
pandas
numpy
rich
scikit-image
scikit-learn
omegaconf
segment-anything; python_version >= "3.10" # (필요 시 별도 설치 소스 변경 가능)
